Move logging into sublibrary
