+++ src/isle/cpp/action/hubbardFermiAction.cpp
@@ -1,7 +1,7 @@
 #include "hubbardFermiAction.hpp"
 
 #include "../core.hpp"
-#include "../bind/logging.hpp"
+#include "../logging/logging.hpp"
 
 using namespace std::complex_literals;
